Move the spec finder under a new namespace, interpreter._import.
